Cod/pyproject.toml
[project]
name = "cod"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.53.0",
]
